Use linear strategy for builders
With Ansible 2.7, `get_mount_facts` would timeout when the entirety of the builder
containers are being spawned concurrently by the builder role. It was
not predictable or repeatable pattern, generally passing on local
machine but failing in CI.

While this slows down the build process, this will ensure the
`get_mount_facts` operation does not time out on package builds.